test(): refine e2e tests
