+++ package.json
@@ -11,10 +11,7 @@
     "prepare": "lerna bootstrap && husky install",
     "start": "lerna run start --scope \"@next-core/brick-container\"",
     "serve": "brick-container-serve",
-    "serve:ci": "concurrently -n 1,2,3 \"npm run serve:ci:1\" \"npm run serve:ci:2\" \"npm run serve:ci:3\"",
-    "serve:ci:1": "brick-container-serve --subdir --no-live-reload --standalone",
-    "serve:ci:2": "npm run serve:ci:1 -- --port=8082 --standalone-micro-apps",
-    "serve:ci:3": "npm run serve:ci:1 -- --port=8083 --standalone-micro-apps --standalone-app-dir='home/'",
+    "serve:e2e": "cross-env NODE_ENV='e2e-test' brick-container-serve",
     "build": "lerna run build",
     "docs:build": "lerna run docs:build",
     "docs:dev": "lerna run docs:dev",
@@ -27,11 +24,8 @@
     "cypress:open": "cypress open",
     "cypress:run": "cypress run --record",
     "cypress:run:no-record": "cypress run",
-    "test:e2e:ci": "npm run test:e2e:ci:1 && npm run test:e2e:ci:2 && npm run test:e2e:ci:3",
-    "test:e2e:ci:1": "npx wait-on -r http://localhost:8081/next/ && npx start-server-and-test serve:ci:1 http://localhost:8081/next/ cypress:run",
-    "test:e2e:ci:2": "npx wait-on -r http://localhost:8082/next/ && npx start-server-and-test serve:ci:2 http://localhost:8082/next/ 'cross-env CYPRESS_BASE_URL=http://localhost:8082 npm run cypress:run'",
-    "test:e2e:ci:3": "npx wait-on -r http://localhost:8083/next/ && npx start-server-and-test serve:ci:3 http://localhost:8083/next/ 'cross-env CYPRESS_BASE_URL=http://localhost:8083 npm run cypress:run'",
-    "test:e2e:no-record": "npx wait-on -r http://localhost:8081/next/ && npx start-server-and-test serve:ci:1 http://localhost:8081/next/ cypress:run:no-record",
+    "test:e2e:ci": "npx wait-on -r http://localhost:8081/next/ && npx start-server-and-test serve:e2e http://localhost:8081/next/ cypress:run",
+    "test:e2e:no-record": "npx wait-on -r http://localhost:8081/next/ && npx start-server-and-test serve:e2e http://localhost:8081/next/ cypress:run:no-record",
     "test:e2e:docker": "docker-compose run --user $(id -u):$(id -g) e2e",
     "lint-staged": "lint-staged"
   },
